requirements.txt
# AI Framework
pydantic-ai==0.0.14
pydantic==2.10.3
openai==1.57.2

# LangGraph for conversation orchestration
langgraph>=0.2.0
langchain-core>=0.3.0

# HTTP and utilities
httpx==0.28.1
python-dotenv==1.0.1
